feat: use new sandbox layout
