+++ packages/analytics/dashboard-renderer/package.json
@@ -40,6 +40,7 @@
   "devDependencies": {
     "@kong-ui-public/analytics-chart": "^0.15.3",
     "@kong-ui-public/i18n": "workspace:^",
+    "@kong-ui-public/sandbox-layout": "workspace:^",
     "@kong/design-tokens": "^1.11.9",
     "@kong/kongponents": "^8.126.2",
     "json-schema-to-ts": "^2.9.2",
